+++ src/Debug/Debug.cpp
@@ -2,5 +2,13 @@
 #include <iostream>
 
 void spic::Debug::LogWarning(const std::string &warning) {
-    std::cerr << "ERROR: " << warning << std::endl;
+    std::cerr << "WARNING: " << warning << std::endl;
+}
+
+void spic::Debug::LogError(const std::string &error) {
+    std::cerr << "ERROR: " << error << std::endl;
+}
+
+void spic::Debug::Log(const std::string &message) {
+    std::cerr << "LOG: " << message << std::endl;
 }